[FIX] 예외처리 추가 : 이미 좋아요한 게시글 체크, 북마크 권한 체크, 피드 상세 위치 정보 존재여부 체크 gitIgnore 내용 추가 : application.yml 파일 보호 처리

+++ src/main/java/com/example/trip/service/FeedService.java
@@ -1,8 +1,6 @@
 package com.example.trip.service;
 
-import com.example.trip.advice.exception.AuthFeedCommentNotFoundException;
-import com.example.trip.advice.exception.AuthFeedNotFoundException;
-import com.example.trip.advice.exception.AuthLikesNotFoundException;
+import com.example.trip.advice.exception.*;
 import com.example.trip.domain.*;
 import com.example.trip.dto.request.FeedRequestDto;
 import com.example.trip.repository.LikeRepository;
@@ -133,7 +131,16 @@ public class FeedService {
     }
 
     public void likeFeed(Long feedDetailLocId, User user) {
-        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId).orElseThrow(() -> new NullPointerException("해당 값이 없습니다."));
+        // 해당 피드 상세 위치 값이 있는지 체크
+        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId)
+                .orElseThrow(() -> new FeedDetailLocNotFoundException());
+
+        //이미 좋아요한 게시글인지 체크
+        List<Likes> didYouLike = likeRepository.findByFeedDetailLocIdAndUserId(feedDetailLocId, user.getId());
+        if (!didYouLike.isEmpty()){
+            throw new AlreadyLikeException();
+        }
+
         Likes like = Likes.builder()
                 .feedDetailLoc(feedDetailLoc)
                 .user(user)
@@ -154,7 +161,9 @@ public class FeedService {
     }
 
     public void bookmarkFeed(Long feedDetailLocId, User user) {
-        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId).orElseThrow(() -> new NullPointerException("해당 값이 없습니다."));
+        // 해당 피드 상세 위치 값이 있는지 체크
+        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId)
+                .orElseThrow(() -> new FeedDetailLocNotFoundException());
         BookMark bookmark = BookMark.builder()
                 .feedDetailLoc(feedDetailLoc)
                 .user(user)
@@ -164,11 +173,19 @@ public class FeedService {
     }
 
     public void unbookmarkFeed(Long feedDetailLocId, User user) {
+        // 북마크를 한 사람만 권한이 있어야함
+        List<BookMark> myBookmark = bookMarkRepository.findByFeedDetailLocIdAndUserId(feedDetailLocId, user.getId());
+
+        if (myBookmark.isEmpty()) {
+            throw new AuthBookMarkNotFoundException();
+        }
         bookMarkRepository.deleteBookmarkFeed(feedDetailLocId, user.getId());
     }
 
     public void registerFeedComment(User user, Long feedDetailLocId, FeedRequestDto.FeedRequestCommentRegisterDto feedRequestCommentRegisterDto) {
-        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId).orElseThrow(() -> new NullPointerException("해당 값이 없습니다."));
+        // 해당 피드 상세 위치 값이 있는지 체크
+        FeedDetailLoc feedDetailLoc = feedDetailLocRepository.findById(feedDetailLocId)
+                .orElseThrow(() -> new FeedDetailLocNotFoundException());
         FeedComment feedComment = FeedComment.builder()
                 .feedDetailLoc(feedDetailLoc)
                 .user(user)
